refactor: reorganize init service package
