+++ src/test/java/com/example/spinlog/integration/init/GenderStatisticsCacheStartupService.java
@@ -1,11 +1,9 @@
-package com.example.spinlog.statistics.scheduled.startup;
+package com.example.spinlog.integration.init;
 
 import com.example.spinlog.global.cache.HashCacheService;
 import com.example.spinlog.statistics.service.fetch.GenderStatisticsRepositoryFetchService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.context.event.ApplicationReadyEvent;
-import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,21 +18,16 @@ import static com.example.spinlog.utils.StatisticsCacheUtils.*;
 @Transactional(readOnly = true) // todo 범위 좁히기
 @RequiredArgsConstructor
 @Slf4j
-class GenderStatisticsStartupService {
+public class GenderStatisticsCacheStartupService {
     private final HashCacheService hashCacheService;
     private final GenderStatisticsRepositoryFetchService genderStatisticsRepositoryFetchService;
 
-    @EventListener(ApplicationReadyEvent.class)
     public void initGenderStatisticsCache() {
         log.info("Start initializing Caching to Redis");
         LocalDate endDate = LocalDate.now();
         LocalDate startDate = endDate.minusDays(PERIOD_CRITERIA);
 
         AllStatisticsMap allData = genderStatisticsRepositoryFetchService.getGenderStatisticsAllData(startDate, endDate);
-
-        // todo 기존 데이터 검증 후 캐싱
-        // todo 이거 필요 없지 않나??? - 어차피 @Scheduled로 주기적으로 검증해준다.
-        // todo scheduled verify method 작성 후 삭제
 
         hashCacheService.putAllDataInHash(
                 getGenderEmotionStatisticsAmountSumKeyName(SPEND),
